+++ dubbo-samples-basic/src/main/java/org/apache/dubbo/samples/basic/BasicConsumer.java
@@ -17,18 +17,31 @@
  *
  */
 
-package com.alibaba.dubbo.samples.basic;
+package org.apache.dubbo.samples.basic;
+
+import org.apache.dubbo.samples.basic.api.DemoService;
 
 import org.springframework.context.support.ClassPathXmlApplicationContext;
 
-public class BasicProvider {
+public class BasicConsumer {
 
-    public static void main(String[] args) throws Exception {
-        new EmbeddedZooKeeper(2181, false).start();
-        ClassPathXmlApplicationContext context = new ClassPathXmlApplicationContext(new String[]{"spring/dubbo-demo-provider.xml"});
+    public static void main(String[] args) {
+        ClassPathXmlApplicationContext context = new ClassPathXmlApplicationContext(new String[]{"spring/dubbo-demo-consumer.xml"});
         context.start();
+        DemoService demoService = (DemoService) context.getBean("demoService"); // get remote service proxy
 
-        System.in.read(); // press any key to exit
+        while (true) {
+            try {
+                Thread.sleep(1000);
+                String hello = demoService.sayHello("world"); // call remote method
+                System.out.println(hello); // get result
+
+            } catch (Throwable throwable) {
+                throwable.printStackTrace();
+            }
+
+
+        }
+
     }
-
 }
