+++ dubbo-samples-basic/src/main/java/org/apache/dubbo/samples/basic/EmbeddedZooKeeper.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.alibaba.dubbo.samples.basic;
+package org.apache.dubbo.samples.basic;
 
 import java.io.File;
 import java.lang.reflect.Method;
