+++ dubbo-samples-basic/src/main/java/org/apache/dubbo/samples/basic/impl/DemoServiceImpl.java
@@ -17,10 +17,22 @@
  *
  */
 
-package com.alibaba.dubbo.samples.basic.api;
+package org.apache.dubbo.samples.basic.impl;
 
-public interface DemoService {
+import org.apache.dubbo.samples.basic.api.DemoService;
 
-    String sayHello(String name);
+import org.apache.dubbo.rpc.RpcContext;
+
+import java.text.SimpleDateFormat;
+import java.util.Date;
+
+public class DemoServiceImpl implements DemoService {
+
+    @Override
+    public String sayHello(String name) {
+        System.out.println("[" + new SimpleDateFormat("HH:mm:ss").format(new Date()) + "] Hello " + name + ", request from consumer: " + RpcContext
+            .getContext().getRemoteAddress());
+        return "Hello " + name + ", response from provider: " + RpcContext.getContext().getLocalAddress();
+    }
 
 }
